+++ plom/client/viewers.py
@@ -210,6 +210,15 @@ class SelectTestQuestion(QDialog):
 
 
 class SolutionViewer(QWidget):
+    """A non-modal dialog for displaying solutions.
+
+    A note on modality: because the super init call has no parent
+    reference, this is a new top-level window (not formally parented
+    by the Annotator or Marker windows).  At least in the Gnome
+    environment, that means it does not stay on top of the Annotator
+    window (unlike for example `QVHistogram` in Manager).
+    """
+
     def __init__(self, parent, fname):
         super().__init__()
         self._annotr = parent