+++ plom/finishMessenger.py
@@ -1,5 +1,5 @@
 # SPDX-License-Identifier: AGPL-3.0-or-later
-# Copyright (C) 2020-2021 Andrew Rechnitzer
+# Copyright (C) 2020-2022 Andrew Rechnitzer
 # Copyright (C) 2020-2022 Colin B. Macdonald
 
 import requests
@@ -55,7 +55,7 @@ class FinishMessenger(BaseMessenger):
         finally:
             self.SRmutex.release()
 
-    def RgetDanglingPages(self):
+    def getDanglingPages(self):
         self.SRmutex.acquire()
         try:
             response = self.get(
@@ -170,3 +170,17 @@ class FinishMessenger(BaseMessenger):
             raise PlomSeriousException(f"Some other sort of error {e}") from None
         finally:
             self.SRmutex.release()
+
+    def getFilesInAllTests(self):
+        with self.SRmutex:
+            try:
+                response = self.get(
+                    "/REP/filesInAllTests",
+                    json={"user": self.user, "token": self.token},
+                )
+                response.raise_for_status()
+                return response.json()
+            except requests.HTTPError as e:
+                if response.status_code in (401, 403):
+                    raise PlomAuthenticationException(response.reason) from None
+                raise PlomSeriousException(f"Some other sort of error {e}") from None