+++ plom/scan/__main__.py
@@ -35,12 +35,9 @@ Overview of the scanning process
           page).  Force the upload these if you really need to; the
           manager will then have to look at them.
 
-  4. Run "plom-scan status" to get a brief summary of scanning to date.
+  4. Run "plom-scan status" to get a summary of scanning to date.
 
-  5. Run "plom-scan bundles" to get a list of the bundles that have been
-          uploaded into the system.
-
-  6. If something goes wrong such as crashes or interruptions, you may
+  5. If something goes wrong such as crashes or interruptions, you may
      need to clear the "scanner" login with the `clear` command.
 
   These steps may be repeated as new PDF files come in: it is not
@@ -54,7 +51,6 @@ __license__ = "AGPL-3.0-or-later"
 
 import argparse
 import os
-from pathlib import Path
 
 from stdiomask import getpass
 
@@ -62,7 +58,6 @@ from plom.scan import __version__
 from plom.scan import clear_login
 from plom.scan import check_and_print_scan_status
 from plom.scan import processScans, uploadImages
-from plom.scan import print_bundle_list
 
 
 def get_parser():
@@ -89,17 +84,20 @@ def get_parser():
     spS = sub.add_parser(
         "status",
         help="Get scanning status report from server",
-        description="Get scanning status report from server.",
+        description="""
+            Get scanning status report from server.
+            You can customize the report using the switches below
+            or omit all switches to get the full report.
+        """,
     )
+    spS.add_argument("--papers", action="store_true", help="show paper info")
+    spS.add_argument("--unknowns", action="store_true", help="Show info about unknowns")
+    spS.add_argument("--bundles", action="store_true", help="Show bundle info")
+
     spC = sub.add_parser(
         "clear",
         help='Clear "scanner" login',
         description='Clear "scanner" login after a crash or other expected event.',
-    )
-    spB = sub.add_parser(
-        "bundles",
-        help="Get a list of bundles in the plom database",
-        description="Get a list of bundles in the plom database.",
     )
     # TODO: maybe in the future?
     # spA = sub.add_parser(
@@ -177,7 +175,7 @@ def get_parser():
         help='Upload "collisions", pages which appear to already be on the server. '
         + "You should not need this option except under exceptional circumstances.",
     )
-    for x in (spU, spS, spC, spB, spP):
+    for x in (spU, spS, spC, spP):
         x.add_argument("-s", "--server", metavar="SERVER[:PORT]", action="store")
         x.add_argument("-w", "--password", type=str, help='for the "scanner" user')
 
@@ -213,9 +211,12 @@ def main():
             msgr=(args.server, args.password),
         )
     elif args.command == "status":
-        check_and_print_scan_status(msgr=(args.server, args.password))
-    elif args.command == "bundles":
-        print_bundle_list(msgr=(args.server, args.password))
+        check_and_print_scan_status(
+            args.papers,
+            args.unknowns,
+            args.bundles,
+            msgr=(args.server, args.password),
+        )
     elif args.command == "clear":
         clear_login(args.server, args.password)
     else: