+++ plom/server/plomServer/routesUpload.py
@@ -449,17 +449,16 @@ class UploadHandler:
         rval = self.server.getUnknownPages()
         return web.json_response(rval, status=200)
 
-    async def getDiscardNames(self, request):
+    async def getDiscardedPages(self, request):
         data = await request.json()
         if not validate_required_fields(data, ["user", "token"]):
             return web.Response(status=400)
         if not self.server.validate(data["user"], data["token"]):
             return web.Response(status=401)
-        if not data["user"] == "manager":
-            return web.Response(status=401)
-
-        rval = self.server.getDiscardNames()
-        return web.json_response(rval, status=200)  # all fine
+        if data["user"] != "manager":
+            raise web.HTTPForbidden(reason="I only speak to the manager")
+        rval = self.server.getDiscardedPages()
+        return web.json_response(rval, status=200)
 
     async def getCollidingPageNames(self, request):
         data = await request.json()
@@ -531,21 +530,6 @@ class UploadHandler:
         rownames = ("pagename", "md5", "orientation", "id", "server_path")
         pagedata = [{k: v for k, v in zip(rownames, val)}]
         return web.json_response(pagedata, status=200)
-
-    async def getDiscardImage(self, request):
-        data = await request.json()
-        if not validate_required_fields(data, ["user", "token", "fileName"]):
-            return web.Response(status=400)
-        if not self.server.validate(data["user"], data["token"]):
-            return web.Response(status=401)
-        if not data["user"] == "manager":
-            return web.Response(status=401)
-
-        rval = self.server.getDiscardImage(data["fileName"])
-        if rval[0]:
-            return web.FileResponse(rval[1], status=200)  # all fine
-        else:
-            return web.Response(status=404)
 
     async def getCollidingImage(self, request):
         data = await request.json()
@@ -984,9 +968,8 @@ class UploadHandler:
         router.add_get("/admin/scannedHWPage", self.getHWPageImage)
         router.add_get("/admin/scannedEXPage", self.getEXPageImage)
         router.add_get("/admin/unknownPages", self.getUnknownPages)
-        router.add_get("/admin/discardNames", self.getDiscardNames)
+        router.add_get("/admin/discardedPages", self.getDiscardedPages)
         router.add_get("/admin/collidingPageNames", self.getCollidingPageNames)
-        router.add_get("/admin/discardImage", self.getDiscardImage)
         router.add_get("/admin/collidingImage", self.getCollidingImage)
         router.add_get("/admin/checkTPage", self.checkTPage)
         router.add_delete("/admin/unknownImage", self.removeUnknownImage)