+++ plom/server/plomServer/serverUpload.py
@@ -272,10 +272,6 @@ def getEXPageImage(self, testNumber, question, order):
     return self.DB.getEXPageImage(testNumber, question, order)
 
 
-def getDiscardImage(self, fname):
-    return self.DB.getDiscardImage(fname)
-
-
 def getCollidingImage(self, fname):
     return self.DB.getCollidingImage(fname)
 
@@ -284,8 +280,8 @@ def getUnknownPages(self):
     return self.DB.getUnknownPages()
 
 
-def getDiscardNames(self):
-    return self.DB.getDiscardNames()
+def getDiscardedPages(self):
+    return self.DB.getDiscardedPages()
 
 
 def getCollidingPageNames(self):