+++ plom/server/theServer.py
@@ -124,10 +124,9 @@ class Server:
         removeAllScannedPages,
         removeSinglePage,
         getUnknownPages,
-        getDiscardNames,
+        getDiscardedPages,
         getCollidingPageNames,
         getCollidingImage,
-        getDiscardImage,
         getTPageImage,
         getHWPageImage,
         getEXPageImage,
@@ -198,7 +197,7 @@ class Server:
         RgetUnusedTests,
         RgetScannedTests,
         RgetIncompleteTests,
-        RgetDanglingPages,
+        getDanglingPages,
         RgetCompleteHW,
         RgetMissingHWQ,
         RgetProgress,
@@ -216,6 +215,7 @@ class Server:
         RgetIDReview,
         RgetUserList,
         RgetUserDetails,
+        getFilesInAllTests,
     )
     from .plomServer.serverSolution import (
         uploadSolutionImage,