+++ CHANGELOG.md
@@ -11,7 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * plom-server will autogenerate a manager password if it is started without one.
 * Manager displays bundle name/page of unknown pages.
 * `plom-scan status` now displays information about unknown pages.
+* `plom-scan status` now displays information about bundles and unknown pages.
+* `plom-finish audit` produces a JSON file for post-grading checking/followup/etc
 * Annotator now has a crop tool (in menu) that allows user to excluded unwanted parts of page. Works with undo/rego.
+
 
 ### Changed
 * Plom now requires Python 3.7.